api-dec/requirements.txt
aiofiles==0.6.0
aniso8601==7.0.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.11.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
promise==2.3
PyJWT==1.7.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.0
Rx==1.6.1
six==1.15.0
starlette==0.14.1
urllib3==1.26.2
uvicorn==0.12.3
